requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets
peft
accelerate
scikit-learn
tqdm
pymupdf
pytesseract
Pillow
opencv-python
ollama
fastapi
uvicorn
python-multipart
requests 